Dependencies added to allow JWT tokens, Changes in the database for user_profile, country, state, address and social_network. Services added for the creation of a new base user and th login. JWT configuration added.

+++ user_service/src/main/java/com/impuls/user_service/model/gateway/SocialNetworkRepository.java
@@ -3,5 +3,5 @@ package com.impuls.user_service.model.gateway;
 import com.impuls.user_service.model.SocialNetwork;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SocialNetworkRepository extends JpaRepository<SocialNetwork, Integer> {
+public interface SocialNetworkRepository extends JpaRepository<SocialNetwork, Long> {
 }
